new POC testing constructor behaviour and renamed header
